Move cache to a separate pkg, as part of refactoring
