+++ snapshots/go/pkg/cache/cache.go
@@ -1,6 +1,6 @@
 /* Copyright 2022 Cognite AS */
 
-package bazel
+package cache
 
 import (
 	"context"
