+++ snapshots/go/pkg/cache/grpc_client.go
@@ -1,6 +1,6 @@
 /* Copyright 2022 Cognite AS */
 
-package bazel
+package cache
 
 import (
 	"context"
